improved some stuff
